+++ models/annonce.go
@@ -1,0 +1,11 @@
+package models
+
+import (
+	"github.com/jinzhu/gorm"
+)
+
+type Annonce struct {
+	gorm.Model
+	UserID uint
+	Cats   []Cats `gorm:"foreignKey:CustomReferer"`
+}
