+++ models/roles.go
@@ -1,0 +1,8 @@
+package models
+
+import "github.com/jinzhu/gorm"
+
+type Roles struct {
+	gorm.Model
+	Name string `gorm:"type:varchar(100);not null"`
+}
