+++ models/user.go
@@ -2,7 +2,6 @@ package models
 
 import (
 	"github.com/jinzhu/gorm"
-	"time"
 )
 
 type User struct {
@@ -14,22 +13,4 @@ type User struct {
 	Roles    []Roles `gorm:"many2many:user_languages;"`
 	Annonce  []Annonce
 	Favorite []Annonce
-}
-
-type Annonce struct {
-	gorm.Model
-	UserID uint
-	Cats   []Cats `gorm:"foreignKey:CustomReferer"`
-}
-
-type Cats struct {
-	gorm.Model
-	Name      string `gorm:"type:varchar(100);not null"`
-	Color     string `gorm:"type:varchar(100);not null"`
-	BirthDate *time.Time
-}
-
-type Roles struct {
-	gorm.Model
-	Name string `gorm:"type:varchar(100);not null"`
 }
